Fix bug

+++ src/constant.rs
@@ -1,4 +1,4 @@
 pub const GROUP_NUMS: u32 = 4;
-pub const CHAIN_NUMS: u32 = 16;
+pub const CHAIN_NUMS: u8 = 16;
 pub const PARALLEL_MINING_WORKS: u32 = 16;
 pub const MINING_STEPS: u64 = 100000;
